Refactor: Create DBInterface to separate DB interface logic from UserCreator
